pyproject.toml
[tool.poetry]
name = "shallwe-core"
version = "0.6.1"
description = "Shallwe UA requirements"
authors = ["featurelle and team"]

[tool.poetry.dependencies]
python = "3.10.18"
absl-py = "2.0.0"
asgiref = "3.7.2"
astunparse = "1.6.3"
beautifulsoup4 = "4.12.2"
blinker = "1.7.0"
cachetools = "5.3.2"
certifi = "2023.7.22"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
cryptography = "41.0.5"
deepface = "0.0.81"
defusedxml = "0.7.1"
deprecated = "1.2.14"
dj-rest-auth = "5.0.2"
django = "5.0.1"
django-allauth = "0.58.2"
django-appconf = "1.0.6"
django-cors-headers = "4.3.1"
django-environ = "0.11.2"
django-extensions = "3.2.3"
django-imagekit = "5.0.0"
django-taggit = "5.0.1"
djangorestframework = "3.14.0"
docopt = "0.6.2"
filelock = "3.13.1"
fire = "0.5.0"
flask = "3.0.0"
flatbuffers = "23.5.26"
gast = "0.5.4"
gdown = "4.7.1"
google-auth = "2.26.2"
google-pasta = "0.2.0"
grpcio = "1.60.0"
gunicorn = "21.2.0"
h5py = "3.10.0"
idna = "3.4"
itsdangerous = "2.1.2"
jinja2 = "3.1.3"
libclang = "16.0.6"
markdown = "3.5.2"
markupsafe = "2.1.3"
ml-dtypes = "0.2.0"
mtcnn = "0.1.1"
numpy = "1.26.3"
oauthlib = "3.2.2"
opencv-python = "4.9.0.80"
opt-einsum = "3.3.0"
packaging = "23.2"
pandas = "2.1.4"
pi-heif = "0.14.0"
pilkit = "3.0"
pillow = "10.2.0"
pipreqs = "0.4.13"
protobuf = "4.23.4"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pyjwt = "2.8.0"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
python3-openid = "3.2.0"
pytz = "2023.3.post1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
retina-face = "0.0.13"
rsa = "4.9"
six = "1.16.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
tensorboard = "2.15.1"
tensorboard-data-server = "0.7.2"
tensorflow-estimator = "2.15.0"
tensorflow-io-gcs-filesystem = "0.35.0"
termcolor = "2.4.0"
tqdm = "4.66.1"
typing-extensions = ">=4.12.2, <4.13"
tzdata = "2023.4"
urllib3 = "2.1.0"
werkzeug = "3.0.1"
wrapt = "1.14.1"
yarg = "0.1.9"
django-nested-admin = "^4.1.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
